Set pixel_modifier with a function
The "magic value" was due to our data being 200x200, or 40,000 pixels. 40,000 * 0.00001 is a value less than one, which was necessary to not influence sum difference. In cases such as the Mandlebrot Set we were running over 2 million pixels so it is possible it affected the outcome, so a function was made to correctly calculate the modifier.